Develop

+++ Ettad.Workflow.Service/Interface/IWorkflowApprovalService.cs
@@ -23,13 +23,13 @@ namespace Ettad.Workflows.Service.Interface
         Task<WorkflowApprovalStepDto> RejectAsync(ApproveRejectWorkflowApprovalDto dto);
         Task<WorkflowApprovalStepDto> ApproveOrReject(ApproveRejectWorkflowApprovalDto dto);
         Task<WorkflowApprovalStepDto> ApproveOrReject(ApproveRejectWorkflowApprovalDto dto, List<IFormFile> files);
-        Task<WorkflowApprovalStep> GetCurrentApprovalStepByRequestIdAsync(int requestId);
+        Task<WorkflowApprovalStep> GetCurrentApprovalStepByRequestIdAsync(long requestId);
         Task<APIOperationResponse<bool>> ProcessActionAsync(ApproveRejectWorkflowApprovalDto model);
         Task<APIOperationResponse<bool>> ProcessActionAsync(ApproveRejectWorkflowApprovalDto model, List<IFormFile> files);
         Task<bool> StartWorkflowAsync(long orderId, WorkflowType workflowType);
         Task<IEnumerable<BaseRequestDto>> GetAllBaseRequestsAsync();
         Task<BaseRequestDto> GetBaseRequestByIdAsync(long requestId);
-        Task<IEnumerable<WorkflowStepDto>> GetPreviousWorkflowStepsForReturn(int requestId);
+        Task<IEnumerable<WorkflowStepDto>> GetPreviousWorkflowStepsForReturn(long requestId);
 
     }
 }
